Initial commit
This is version 1.0 of aLOTofImaginArches, an open-source Matlab app for the graphical analysis of masonry arches. The software is intended to be an open-source Matlab educational tool that facilitates interactive demonstrations, engaging users in experimentation with the concept of funicular solutions. It is designed for master’s students in the MHMS and HMWS courses and is particularly useful for illustrating the concept of infinite (to the power of 3!) funicular solutions for masonry arches.